naukri-agent/requirements.txt
selenium
webdriver-manager
python-dotenv
pytest
openai
google-generativeai
PyPDF2
jinja2
